pyproject.toml
[tool.poetry]
name = "stability-radius"
version = "0.1.0"
description = "Stability radii for power systems"
authors = ["Ygor"]

[tool.poetry.dependencies]
python = "^3.10"
pandapower = "^2.13"          
numpy = "^1.26"
scipy = "^1.11"
pandas = "^2.2"
matplotlib = "^3.8"
hydra-core = "^1.3"           
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"